+++ src/components/kamea/index.ts
@@ -1,0 +1,2 @@
+export { Kamea } from './Kamea';
+export { KameaCanvas } from './KameaCanvas';